Make logic to invalidate old transactions. Use efficient data structures (TreeMaps are always sorted).

+++ src/main/java/com/statistics/domain/IStatisticsService.java
@@ -3,7 +3,7 @@ package com.statistics.domain;
 public interface IStatisticsService {
 
 	void addTransaction(Transaction transaction);
-	int getStatisticsCount();
+	long getStatisticsCount();
 	void reset();
 	StatisticsResult getStatisticsResult();
